docs(deps): add NVASM download link
Add README with a download link for acquiring NVASM, Nvidia's vertex and
pixel shader macro assembler, for development purposes. NVASM is needed
to compile .nvp and .nvv shader files used by the game but cannot be
included in the repository due to its proprietary license, which is not
compatible with GPLv3.

This dependency should be removed or replaced with a GPLv3-compatible
alternative before release. In the meantime, it can be acquired from the
following link for use during development.

Acquired-from: https://archive.org/download/nvasm_1_42/nvasm_1_42.zip
License: proprietary (see README)